+++ Assignments/assignment6/src/lib.rs
@@ -1,6 +1,3 @@
 pub mod ip_configuration;
 pub mod quadrant;
-pub mod test {
-    pub mod ip_configuration_test;
-    pub mod quadrant_test;
-}
+pub mod test;
